Now version is a interface{}
In order to allow wrap any external command and allow any kind
of version format, including valid json, we must model the
version as interface{}.

If the version is a pure string, we pass it as a string. If it is
a JSON it will be passed as a JSON struct.

json.RawMessage did not work, as encoding and decoding the same
json would result in a encoded binary of the json.RawMessage.