Clear separation of smuggler params from other
Refactor the models to separate the parameters defined
for smuggler under the key smuggler_params from any other
params defined in the source or in the task.

This allows create a filtered version of the request
without the smuggler specific params.

The ResourceRequest will include:
 * The smuggler and the extra params in their own attributes.
 * The original raw request ready to serialise
 * The filtered request without smuggler config to serialise

Also refactor the fixtures that read from a YAML pipeline fixture, so
that it creates the request in JSON as it would be created in
concourse atc.