Simplify command syntax and one liners
Now the configuration supports define the commands as strings or
as commands definitions with path and args.

Solved by using interface{} and evaluating it during runtime.

We also represent the ands a hash like `in: val`.
This will simplify the merge of the local config file ‘smuggler.yml’
with the incoming request, as it is simply merge two hashes.

Update README for MVP version