Use mergo to merge `smuggler.yml` with request
We use the library github.com/imdario/mergo in order to merge the 
local config file `smuggler.yml` with the source of the request.

To do so, we parse both the config file and the request as 
`interface{}`, wrap the config under a hash map with 
`{"source": val}` so they match structures and merge the request
overriding on the config. Finally we covert the result back
to a json string, so that we can use the constructor of 
ResourceRequest to parse it.