Use docker multistage builds to build smuggler
We want to simplify the creation of the docker containers by using
the new docker feature multistage builds[1]

In one stage we will build the smuggler binary, or any other binary
and then copy it to the next layers.

We also get rid of all the unnecessary boilerplate and scripts.