Accept versions that are valid json as raw string
When writing to ${SMUGGLER_OUTPUT_DIR}/versions, each line will
be parsed to check if it is a json with a type map[string]string.
If it is so, it will be directly used as the version struct,
otherwise the value would be wrapped in a hash: { "ID": "$val" }

But if the string is a valid json, but with a different type,
smuggler fails due a type missmatch.

That behaviour is incorrect. the line should the threated
as a raw string and also sbe wrapped in the hash.

In this commit NewVersion always success, and only decodes it as a
json if the line written in ${SMUGGLER_OUTPUT_DIR}/versions has
the format map[string]string, like:

    {
         "ref": "aaa",
         "branch": "master"
    }

There is no reason to return a error code, so we change the
signature and usages.